requirements.txt
openai
streamlit
langchain
PyPDF2
langchain-community
python-dotenv
